backend/pyproject.toml
[tool.poetry]
name = "foodgram-project-react"
version = "0.1.0"
description = ""
authors = ["khalaimovda <halaimov_15@mail.ru>"]
[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
drf-yasg = "^1.20.0"
drf-extra-fields = "^3.2.1"
python-dotenv = "^0.19.2"
django-filter = "^21.1"
reportlab = "^3.6.6"
django-colorfield = "^0.6.3"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pep8-naming = "^0.12.1"
flake8-broken-line = "^0.4.0"
flake8-return = "^1.1.3"
flake8-isort = "^4.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
